% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0_filePath.R
\docType{methods}
\name{baseName}
\alias{baseName}
\title{Basename of a FilePath object}
\usage{
\S4method{baseName}{FilePath}(fp)
}
\arguments{
\item{fp}{}
}
\value{
character vector of basenames
}
\description{
Basename of a FilePath object
}
\examples{
fp = filePath(R.home())
baseName(fp)
}
\author{
kaiyin
}

