% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0_readFactory.R
\name{bim_fam_info}
\alias{bimInfo}
\alias{bim_fam_info}
\alias{famInfo}
\title{.bim and .fam file information functions}
\usage{
bimInfo(filename)

famInfo(filename)
}
\description{
These two functions \code{bimInfo, famInfo} are generated by a
factory function. They are similar to readInfo, only with
pre-determined colnames (since they are really known).
}

