% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0_nonExistentFiles.R
\name{checkFileExist}
\alias{checkFileExist}
\title{Stop when any file does not exist}
\usage{
checkFileExist(files)
}
\arguments{
\item{files}{character vector. File paths you want to check.}
}
\description{
Stop when any file does not exist
}
\examples{
\\donotrun{
checkFileExist(R.home())
checkFileExist(sapply(1:5, function(i) tempfile()))
checkFileExist(sapply(1:5, function(i) tempdir()))
}
}
\author{
kaiyin
}

