% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0_df2ffdf.R
\name{df2ffdf}
\alias{df2ffdf}
\title{Converts a data frame to an ffdf}
\usage{
df2ffdf(df, ...)
}
\arguments{
\item{df}{the data.frame you want to convert.}
}
\value{
an ffdf object.
}
\description{
Converts a data frame to an ffdf
}
\examples{
require(ff)
d = data.frame(x = rnorm(10), y = sapply(65:74, function(i) rawToChar(as.raw(i))))
d.ffd = df2ffdf(d)
class(d.ffd)
d1 = ff::as.ram(d.ffd)
all(d1 == d) # true
}

