% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0_filePath.R
\docType{methods}
\name{dirName}
\alias{dirName}
\title{Directory name of a file path}
\usage{
\S4method{dirName}{FilePath}(fp)
}
\arguments{
\item{fp}{FilePath object}
}
\value{
character vector of directories
}
\description{
Directory name of a file path
}
\examples{
fp = filePath(R.home())
dirName(fp)
}
\author{
kaiyin
}

