% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0_filePath.R
\name{filePath}
\alias{filePath}
\title{Constructor for FilePath class}
\usage{
filePath(s)
}
\arguments{
\item{s}{character, path to file or dir}
}
\value{
FilePath object
}
\description{
Constructor for FilePath class
}
\author{
kaiyin
}

