% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0_ifLen.R
\name{ifLen}
\alias{ifLen}
\title{IfLen macro}
\usage{
ifLen(df = stop("df not supplied"), body1 = stop("body1 not supplied"),
  body2 = { })
}
\arguments{
\item{df}{An object which can be passed to \code{length}}

\item{body1}{If \code{length(df)} is not zero, then this clause is evaluated, otherwise, body2 is evaluated.}

\item{body2}{See above.}
}
\description{
Check whether a object has non-zero length, and
eval expression accordingly.
}
\examples{
ifLen(c(1, 2), { print('yes!') }, {print("no!")})
}
\author{
kaiyin
}

