% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0_ifLen.R
\name{ifLetLen}
\alias{ifLetLen}
\title{IfLetLen macro}
\usage{
ifLetLen(sym_str = stop("sym_str not supplied"), x = stop("x not supplied"),
  body1 = stop("body1 not supplied"), body2 = { })
}
\arguments{
\item{x}{the predicate to be evalueated, and to be assigned to a temporary var called \code{..temp..}}

\item{body1}{expression to be evaluated when \code{..temp..} is TRUE.}

\item{body2}{expression to be evaluated when \code{..temp..} is FALSE.}
}
\description{
Similar to ifLet, but conditioned on whether the length of
the result of \code{eval(x)} is 0.
}
\examples{
ifLetLen("..temp..", 1:3, {print(paste("true.", as.character(..temp..)))},
		{print(paste("false.", as.character(..temp..)))})
}
\author{
kaiyin
}

