% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bedinfo.R
\docType{methods}
\name{isSetup,PlInfo-method}
\alias{isSetup,PlInfo-method}
\title{Check if a directory containing .bed .fam and .bim files is properly setup}
\usage{
\S4method{isSetup}{PlInfo}(pl_info)
}
\arguments{
\item{pl_info}{PlInfo object}
}
\value{
TRUE or FALSE
}
\description{
Check if a directory containing .bed .fam and .bim files is properly setup
}
\examples{
# see examples in plInfo
}
\author{
kaiyin
}

