% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0_loadFFDF.R
\docType{methods}
\name{loadFFDF}
\alias{loadFFDF}
\alias{loadFFDF,character-method}
\title{Load ffdf directory}
\usage{
\S4method{loadFFDF}{FilePath}(dir)

\S4method{loadFFDF}{character}(dir)
}
\arguments{
\item{dir}{the ffdf directory to be loaded}
}
\value{
an ffdf object
}
\description{
Load ffdf directory
}
\examples{
\\donotrun{
iris_ff = df2ffdf(iris)
save.ffdf(iris_ff, dir = "iris_ff")
rm(iris_ff)
# error: object "iris_ff" not found
head(iris_ff)
iris_ff = loadFFDF("iris_ff")
# print out head of iris data, in ffdf format
head(iris_ff)
}
}
\author{
kaiyin
}

