% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bedinfo.R
\docType{methods}
\name{loadFrq}
\alias{loadFrq}
\title{Load plink frq file (in ff format) into R}
\arguments{
\item{pl_info}{PlInfo object}
}
\value{
ffdf object
}
\description{
Load plink frq file (in ff format) into R
}
\examples{
# see examples in plInfo
}
\author{
kaiyin
}

