% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bedinfo.R
\docType{methods}
\name{loading_functions}
\alias{loading_functions}
\title{Loading function for .bim .fam and .frq data}
\description{
These three S4 methods \code{loadBim, loadFam, loadFrq} are
generated by a factory function. They all receive an PlInfo object as parameter.
In a properly setup plink directory, .bim .fam and .frq
files should have been converted to ff format. These three
functions load the corresponding ff backing files on disk.
}
\author{
kaiyin
}

