% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0_nonExistentFiles.R
\docType{methods}
\name{nonExistentFiles,character-method}
\alias{nonExistentFiles,character-method}
\title{Non-existent files from a vector of filenames}
\usage{
\S4method{nonExistentFiles}{character}(filenames)
}
\arguments{
\item{filenames}{character A vector of filenames}
}
\value{
A character vector of file paths that do not exist.
}
\description{
This function receives a vector of filenames as parameter,
and returns a vector of non-existent files among them.
}
\examples{
\\donotrun{
nonExistentFiles(R.home())
nonExistentFiles(sapply(1:5, function(i) tempfile()))
nonExistentFiles(sapply(1:5, function(i) tempdir()))
}
}
\author{
kaiyin
}

