% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bedinfo.R
\name{plInfo}
\alias{plInfo}
\title{Info about plink bed files}
\usage{
plInfo(bedstem)
}
\arguments{
\item{bedstem}{path of bed file excluding extension name}
}
\value{
a PlInfo object
}
\description{
Info about plink bed files, including the root directory,
paths of plink .bed, .bim, .fam and .frq files, ff backing
directories for .bim, .fam and .frq files. This function is a
constructor for PlInfo s4 class.
}
\examples{
\\donotrun{
pl_info = plInfo("/Users/kaiyin/EclipseWorkspace/collapsabel2/tests/testthat/mmp13")
isSetup(pl_info) # false
setup(pl_info)
isSetup(pl_info) # true
bim_ff = loadMyBim(pl_info)
head(bim_ff)
fam_ff = loadFam(pl_info)
head(fam_ff)
summary(fam_ff[, "IID"])
which(fam_ff[, "IID"] == "10425")
frq_ff = loadFrq(pl_info)
head(frq_ff)
}
}
\author{
kaiyin
}

