% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0_readFactory.R
\name{readBim}
\alias{readBim}
\title{Read plink .bim files}
\usage{
readBim(filename, cn_select)
}
\arguments{
\item{filename}{.bim file path}

\item{cn_select}{a character vector for selected colnames}
}
\value{
a data.frame
}
\description{
Read plink .bim files
}
\examples{
\\donotrun{
bim = readBim("/Users/kaiyin/EclipseWorkspace/collapsabel2/tests/testthat/mmp13.bim",
		cn_select = "..all")
head(bim)
summary(bim)
bim_info = bimInfo("/Users/kaiyin/EclipseWorkspace/collapsabel2/tests/testthat/mmp13.bim")
bim_info@cnames
bim1 = bim_info@read_fun(bim_info, c("CHR", "SNP", "BP"))
head(bim1)
}
}
\author{
kaiyin
}

