% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0_readFactory.R
\name{readFam}
\alias{readFam}
\title{Read plink .fam files}
\usage{
readFam(filename, cn_select)
}
\arguments{
\item{filename}{.fam file path}

\item{cn_select}{a character vector for selected colnames}
}
\value{
a data.frame
}
\description{
Read plink .fam files
}
\examples{
\\donotrun{
fam = readFam("/Users/kaiyin/EclipseWorkspace/collapsabel2/tests/testthat/mmp13.fam", cn_select = "..all")
head(fam)
summary(fam)
fam_info = famInfo("/Users/kaiyin/EclipseWorkspace/collapsabel2/tests/testthat/mmp13.fam")
fam_info@cnames
fam1 = fam_info@read_fun(fam_info, c("FID", "IID"))
head(fam1)
}
}
\author{
kaiyin
}

