% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0_readFactory.R
\docType{methods}
\name{readInfo}
\alias{readInfo}
\alias{readInfo,character,character-method}
\title{ReadInfo constructor}
\usage{
\S4method{readInfo}{character,missing}(filename)

\S4method{readInfo}{character,character}(filename, cnames)
}
\arguments{
\item{filename}{Path of the file to read}
}
\value{
ReadInfo object
}
\description{
This function takes a file path as parameter, assuming the file is whitspace delimited,
not quoted, and has a header line. It returns a ReadInfo object.
}
\examples{
\\donotrun{
ri = readInfo("/Users/kaiyin/EclipseWorkspace/collapsabel2/tests/testthat/mmp13.frq")
getSlots("ReadInfo")
ri@cnames
ri@filename
ri@header
frq = ri@read_fun(ri, cn_select = "..all")
head(frq)
print(ri@cnames)
frq1 = ri@read_fun(ri, ri@cnames[1:4])
head(frq1)
}
}
\author{
kaiyin
}

