% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0_loadFFDF.R
\docType{methods}
\name{saveFFDF}
\alias{saveFFDF}
\alias{saveFFDF,data.frame,character,missing-method}
\title{Save an R data.frame to ff format}
\usage{
\S4method{saveFFDF}{data.frame,character,logical}(df, dir, overwrite)

\S4method{saveFFDF}{data.frame,character,missing}(df, dir, overwrite)
}
\arguments{
\item{df}{data.frame to be saved}

\item{dir}{ff directory}

\item{overwrite}{logical. Default to TRUE}
}
\description{
Save an R data.frame to ff format
}
\examples{
\\donotrun{
saveFFDF(iris, "/tmp/iris")
iris1 = loadFFDF("/tmp/iris")
iris1[1, 1] == 5.1
}
}
\author{
kaiyin
}

