% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bedinfo.R
\docType{methods}
\name{setup,PlInfo-method}
\alias{setup,PlInfo-method}
\title{Setup up a directory containing plink files}
\usage{
\S4method{setup}{PlInfo}(pl_info)
}
\arguments{
\item{pl_info}{}
}
\description{
Setup up a directory containing plink files
}
\examples{
# see examples in plInfo
}
\author{
kaiyin
}

