% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0_strConcat.R
\docType{methods}
\name{strConcat}
\alias{strConcat}
\alias{strConcat,character,missing-method}
\title{Concatenate a vector of strings}
\usage{
\S4method{strConcat}{character,character}(ss, sep)

\S4method{strConcat}{character,missing}(ss, sep)
}
\arguments{
\item{ss}{vector of strings}

\item{sep}{a length-1 string used as separator, default to ""}
}
\value{
a string
}
\description{
Concatenate a vector of strings
}
\examples{
strConcat(letters)
strConcat(letters, " ")
}
\author{
kaiyin
}

